for #2084, add ComplexExpressionSegment
