+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/antlr/sql/segment/dml/expr/simple/LiteralExpressionSegment.java
@@ -15,28 +15,45 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.parse.antlr.sql.segment.dml.expr;
+package org.apache.shardingsphere.core.parse.antlr.sql.segment.dml.expr.simple;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.core.parse.old.lexer.token.DefaultKeyword;
 import org.apache.shardingsphere.core.parse.old.parser.expression.SQLExpression;
 import org.apache.shardingsphere.core.parse.old.parser.expression.SQLIgnoreExpression;
+import org.apache.shardingsphere.core.parse.old.parser.expression.SQLNumberExpression;
+import org.apache.shardingsphere.core.parse.old.parser.expression.SQLTextExpression;
 
 /**
- * Common expression segment.
+ * Literal expression segment.
  * 
+ * @author duhongjun
+ * @author panjuan
  * @author zhangliang
  */
 @RequiredArgsConstructor
 @Getter
-public final class CommonExpressionSegment implements ExpressionSegment {
+public final class LiteralExpressionSegment implements SimpleExpressionSegment {
     
     private final int startIndex;
     
     private final int stopIndex;
     
+    private final Object literals;
+    
     @Override
     public SQLExpression getSQLExpression(final String sql) {
-        return new SQLIgnoreExpression(sql.substring(startIndex, stopIndex + 1));
-    }
+        if (literals instanceof Number) {
+            return new SQLNumberExpression((Number) literals);
+        }
+        if (literals instanceof String) {
+            return new SQLTextExpression(literals.toString());
+        }
+        String value = sql.substring(startIndex, stopIndex + 1);
+        if (DefaultKeyword.NULL.name().equalsIgnoreCase(value)) {
+            return new SQLNumberExpression(null);
+        }
+        return new SQLIgnoreExpression(value);
+    } 
 }
